requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
pdfplumber==0.11.0
pandas==2.2.2
openpyxl==3.1.4
beautifulsoup4==4.12.3
requests==2.32.3
motor==3.4.0
pymongo==4.7.2
apscheduler==3.10.4
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
lxml==5.2.1
Pillow==10.3.0
